requirements.txt
requests
beautifulsoup4
feedgen
lxml
python-dateutil
pytz